Put 22.04 back
I removed builds for 22.04 because I didn't want to publish a build that
would be broken for most people. But based on the number of people
saying it works for them, it's clear that's not the case (even if it's
broken for some people). I think I'll put it back while I continue
debugging the problem.