Install zig in configure step
I can't currently put this binary package in a PPA because the PPA
server builds from source itself. That's a problem because the PPA build
server is missing zig. We can work around this with a little "hack".
Install zig to /opt an symlink to /usr/local/bin as part of the
"configure" step.

Fixes #9