Build with FPM
Distributing the build via a PPA seems difficult/impossible because
ubuntu doesn't have an official version of zig, and it's difficult to
use an unofficial version of zig as part of our build process.

As such, let's pivot. Let's simplify our build by using
[fpm](https://github.com/jordansissel/fpm) instead of debuild, and
produce only a binary .deb package.

We can find alternatives to a PPA to distribute this -- perhaps an apt
repo on GitHub or via another tool like Cloudsmith.