Simplify to dpkg-deb
My first iteration of this used debuild to produce a .deb package, as
recommended in the [Guide for Debian
Maintainers](https://www.debian.org/doc/manuals/debmake-doc/index.en.html).
However, it's become clear that it will be difficult to get a PPA
working because there's not official Ubuntu/Debian zig package, which
makes it difficult to use as part of your tool chain. (PPAs work by
uploading source .deb packages and building binary packages on
launchpad's build servers, which requires zig to be available as an
official package.)

Instead, I'm opting to simplify this build process by using `dpkg-deb`
to build a binary `.deb` package, doing away with the
currently-unnecessary complexity of debuild. I still believe we can
upload this binary .deb package to an apt repository somewhere, just not
a Launchpad PPA.