GitHub Action Workflow for Release
In bf88a72, I ripped out the part of our workflow that pushes release
artifacts as I was experimenting with a matrix build. Let's put that
back now. As a separate build job that depends on the first, so we can
include artifacts from all Ubuntu versions in the release.