Use Docker for our build env
I think this will be a benefit overall, it gives us a reproducible build
environment that can be used in or outside of GitHub Actions.

Also, it will let us build in versions of Ubuntu/Debian other than what
GitHub Actions supports for runners, and possibly even other
architectures with Qemu.