De-Duplicate Ghostty Version from control File
Every time we update the version, we have to update it both in the build
script and in the DEBIAN/control file. Let's make out build script write
the control file as part of the build process so we only have to update
it in one place.

This should also remove a roadblock for doing nightly builds more
easily, since our build script will be the source of truth for the
version.