Move binary package to subdir
I've made some progress toward getting a PPA working and I want to
integrate that to the main branch (especially since GitHub workflows
need to be run from the main branch in some cases). In preparation to do
that, move the binary.deb build to a subdir. The existing binary .deb
build remains the best stable thing to use for the time being; the PPA
is experimental for now.