Build Ghostty & Zig for PPA
Experimented with this on a separate branch for a while, and it's ready
to be merged to the main branch. Doing so as a copy/paste since the
branches don't share history anyway, and this is a reasonable starting
point. Having this on the main branch will make it easier to use GitHub
workflows since for certain things those must exist on or be run from
main.